Extract Prometheus metrics to metrics.go
