+++ .gitignore
@@ -13,7 +13,7 @@
 # Project-local glide cache, RE: https://github.com/Masterminds/glide/issues/736
 .glide/
 
-bank-vaults
+./bank-vaults
 
 # Operator
 operator/tmp/_output/bin/*
