+++ cmd/bank-vaults/unseal.go
@@ -8,10 +8,6 @@ import (
 	"github.com/hashicorp/vault/api"
 	"github.com/sirupsen/logrus"
 	"github.com/spf13/cobra"
-	"github.com/gin-gonic/gin"
-	"github.com/prometheus/client_golang/prometheus"
-
-	"github.com/prometheus/client_golang/prometheus/promhttp"
 )
 
 const cfgUnsealPeriod = "unseal-period"
@@ -25,8 +21,6 @@ type unsealCfg struct {
 }
 
 var unsealConfig unsealCfg
-
-var exporter Exporter
 
 var unsealCmd = &cobra.Command{
 	Use:   "unseal",
@@ -71,7 +65,8 @@ from one of the followings:
 			logrus.Fatalf("error creating vault helper: %s", err.Error())
 		}
 
-		go 	metrics()
+		go metrics.Run()
+
 		for {
 			func() {
 				if unsealConfig.proceedInit {
@@ -91,9 +86,9 @@ from one of the followings:
 					return
 				}
 
-				exporter.Initialized=true
+				metrics.Initialized = true
 				logrus.Infof("vault sealed: %t", sealed)
-				exporter.Sealed=sealed
+				metrics.Sealed = sealed
 
 				// If vault is not sealed, we stop here and wait another unsealPeriod
 				if !sealed {
@@ -121,56 +116,6 @@ func exitIfNecessary(code int) {
 	if unsealConfig.runOnce {
 		os.Exit(code)
 	}
-}
-
-
-const prometheusNS = "vault"
-var (
-	initialized = prometheus.NewDesc(
-		prometheus.BuildFQName(prometheusNS, "", "initialized"),
-		"Is the Vault initialised.",
-		nil, nil,
-	)
-	sealed = prometheus.NewDesc(
-		prometheus.BuildFQName(prometheusNS, "", "sealed"),
-		"Is the Vault node sealed.",
-		nil, nil,
-	)
-)
-
-type Exporter struct {
-	Sealed bool
-	Initialized bool
-}
-func (e *Exporter) Describe(ch chan<- *prometheus.Desc) {
-	ch <- initialized
-	ch <- sealed
-}
-
-func bToF(b bool) float64 {
-	if b {
-		return 1
-	}
-	return 0
-}
-
-func (e *Exporter) Collect(ch chan<- prometheus.Metric) {
-	ch <- prometheus.MustNewConstMetric(
-		sealed, prometheus.GaugeValue, bToF(e.Sealed),
-	)
-	ch <- prometheus.MustNewConstMetric(
-		initialized, prometheus.GaugeValue, bToF(e.Initialized),
-	)
-}
-func metrics(){
-	var defaultMetricsPath = "/metrics"
-	var defaultMetricsPort = ":9091"
-	logrus.Infof("vault metrics exporter enabled: %s%s",defaultMetricsPort, defaultMetricsPath)
-	prometheus.MustRegister(&exporter)
-	server := gin.New()
-	server.Use(gin.Logger(), gin.ErrorLogger())
-	server.GET(defaultMetricsPath, gin.WrapH(promhttp.Handler()))
-	server.Run(defaultMetricsPort)
 }
 
 func init() {
